Refactor and fixed imports
