+++ src/app/team/page.tsx
@@ -48,19 +48,16 @@ export default async function TeamPage({ params }: { params: { teampage: string 
           <span className="font-normal text-gray-200 pr-1">&#60; </span> Welcome <span className="font-jersey_25 text-blue-700 text-3xl">{session.user.name}</span> ! <span className="font-normal pl-1 text-gray-200">/&#62;</span>
         </h1>
         <div>
-          Here you will find the schedules for your rounds and interviews.
-          <br />Please make sure to be on time.
+          Here you will find the schedules for your rounds and interviews as well as results for each round.
+          <br />Please make sure to be on time for rounds and interviews.
           <br />Don&apos;t forget to add the link to your documents down below.
           <br />A link to a google drive folder is fine, but remember we can see the last time it was updated/created. (Ensure permissions are correct for us to access the docs)
         </div>
 
       </div>
 
-      {/* <SwitchButton variant="DOCS" onClick={onClick} /> */}
-
       {team && <TeamInfo team={team} />}
       
-
       <Footer />
     </div>
   );
